Add includeAdded for summary query
Add includeAdded for summary query.  Previously, be default all books
added, regardless of whether they'd been read or time in the specified
time frame were included in the image montage.  This parameter allows
this to be toggled to only read books, but the parameter defaults to
true, so the default behaviour remains the same.