Attempt to prove series lookup in some cases
I found attempting to lookup the culture series by invoking a
same series action on 'Consider Phlebas' resulted in a 'series not
found error'.  After debugging I found the issue was that finito was
saving the books under 'Iain M. Banks', whilst Wikidata was using
'Iain Banks'.  Add a fallback to the Sparql query which handles this with
an OR condition on the first and last part of an author string.